Collected product can be post to/delete from server database

+++ app/src/main/java/app/appworks/school/stylish/data/collected/CollectedFormat.kt
@@ -2,5 +2,5 @@ package app.appworks.school.stylish.data.collected
 
 data class CollectedFormat(
     val userId: Int,
-    val productId: Int
+    val productId: Long
 )
